+++ tokora/src/parser/many/delim/repeated/at_least.rs
@@ -42,18 +42,25 @@ where
   {
     let min = self.parser.parser.minimum().get();
 
-    DelimitedBy::<_, Delim>::new(self.parser.parser.parser_mut()).parse_repeated(
-      inp,
-      &mut self.container,
-      |nums, inp, span| {
-        if min > nums {
-          inp
-            .emitter()
-            .emit_too_few(TooFew::of(span.clone(), nums, min))?;
-        }
+    self
+      .attempt(|c| {
+        let Collect {
+          parser, container, ..
+        } = c;
+        DelimitedBy::<_, Delim>::new(parser.parser.parser_mut()).parse_repeated(
+          inp,
+          container,
+          |nums, inp, span| {
+            if min > nums {
+              inp
+                .emitter()
+                .emit_too_few(TooFew::of(span.clone(), nums, min))?;
+            }
 
-        Ok(())
-      },
-    )
+            Ok(())
+          },
+        )
+      })
+      .map(|(_, collected)| collected)
   }
 }