+++ tokora/src/parser/many/repeated/bounded.rs
@@ -33,9 +33,8 @@ where
     Ctx: ParseContext<'inp, L, Lang>,
   {
     self
-      .as_mut()
-      .parse_input(inp)
-      .map(|_| core::mem::take(&mut self.container))
+      .attempt(|mut c| c.parse_input(inp))
+      .map(|(_, collected)| collected)
   }
 }
 
@@ -70,9 +69,8 @@ where
     Ctx: ParseContext<'inp, L, Lang>,
   {
     self
-      .as_mut()
-      .parse_input(inp)
-      .map(|span| Spanned::new(span, core::mem::take(&mut self.container)))
+      .attempt(|mut c| c.parse_input(inp))
+      .map(|(span, collected)| Spanned::new(span, collected))
   }
 }