+++ tokora/src/parser/many/repeated_while/unbounded.rs
@@ -23,9 +23,8 @@ where
     Ctx: ParseContext<'inp, L, Lang>,
   {
     self
-      .as_mut()
-      .parse_input(inp)
-      .map(|_| core::mem::take(&mut self.container))
+      .attempt(|mut c| c.parse_input(inp))
+      .map(|(_, collected)| collected)
   }
 }
 
@@ -52,9 +51,8 @@ where
     Ctx: ParseContext<'inp, L, Lang>,
   {
     self
-      .as_mut()
-      .parse_input(inp)
-      .map(|span| Spanned::new(span, core::mem::take(&mut self.container)))
+      .attempt(|mut c| c.parse_input(inp))
+      .map(|(span, collected)| Spanned::new(span, collected))
   }
 }